Heavy plug-in module refactoring for retry.
Problem: haskell-cabal sorts before mac-homebrew, and on my Mac systems,
ghc and cabal are installed via homebrew.

Solution: Make a way to retrigger plug-in discovery if it might change
the success or failure.

While I may make this discovery automatic in the future, plug-ins that
know that they import new subsystems (like mac-homebrew) can now call
__hszsh_retry_plugin_search to force rediscovery of plugins that have
failed discovery previously.

Simply renaming directories to enforce an order is not sufficient:
discovery is performed as a batch, and *then* initialization is
performed as a batch.

- Moved major functionality into functions. This primarily eliminates
  the associative array ${__hszsh_current_plugin}; the value
  ${__hszsh_current_plugin[path]} has been replaced with
  ${current_plugin_path} (see git/init).
- This also eliminates ${__hszsh_current_plugin_name} in favour of the
  name ${plugin_detect_file} (see macvim/detect).
- Modified mac-homebrew/init to use __hszsh_retry_plugin_search as the
  cause of this problem.
-