Fixing detection with execution.
Several of my plug-ins were not initializing properly (invisibly so)
because they were looking for `command is-mac` instead of `do is-mac`.
The former always returns true (I always have a function called is-mac)
whereas the latter only returns true if we are on a Mac.

Removed pybugz detection file (this is related to other changes
forthcoming).